+++ lib/src/extensions/list_extensions.dart
@@ -1,3 +1,4 @@
+import 'package:coder_matthews_extensions/src/helpers/extension_helper.dart';
 import 'package:flutter/material.dart';
 import '../helpers/enums.dart';
 import 'object_extensions.dart';
@@ -377,40 +378,11 @@ extension NullableListExtn<T> on Iterable<T>? {
   Iterable<T> orderBy<K>(K? Function(T element) op, [OrderDirection dir = OrderDirection.asc]) {
     if (this == null) return [];
     var shadowThis = this!;
-    int Function(T a, T b) sorter = switch (K) {
-      (num) => (a, b) =>
-          _orderByHelper(op(a), op(b)).$2 ? _orderByHelper(op(a), op(b)).$1! : (op(a) as num).compareTo(op(b) as num),
-      (DateTime) => (a, b) => _orderByHelper(op(a), op(b)).$2
-          ? _orderByHelper(op(a), op(b)).$1!
-          : (op(a) as DateTime).compareTo(op(b) as DateTime),
-      (Duration) => (a, b) => _orderByHelper(op(a), op(b)).$2
-          ? _orderByHelper(op(a), op(b)).$1!
-          : (op(a) as Duration).compareTo(op(b) as Duration),
-      (TimeOfDay) => (a, b) {
-          if (_orderByHelper(op(a), op(b)).$2) {
-            return _orderByHelper(op(a), op(b)).$1!;
-          } else {
-            var aTime = (op(a) as TimeOfDay).hour + (op(a) as TimeOfDay).minute / 60;
-            var bTime = (op(b) as TimeOfDay).hour + (op(b) as TimeOfDay).minute / 60;
-            return aTime.compareTo(bTime);
-          }
-        },
-      (Enum) => (a, b) => _orderByHelper(op(a), op(b)).$2
-          ? _orderByHelper(op(a), op(b)).$1!
-          : (op(a) as Enum).name.compareTo((op(b) as Enum).name),
-      (String) => (a, b) => _orderByHelper(op(a), op(b)).$2
-          ? _orderByHelper(op(a), op(b)).$1!
-          : op(a).toString().compareTo(op(b).toString()),
-      _ => throw UnsupportedError('$K data type is not supported by this order by')
-    };
-    return List.from(shadowThis)..sort((a, b) => sorter(a, b) * (dir == OrderDirection.asc ? 1 : -1));
-  }
-
-  (int?, bool) _orderByHelper<O>(O a, O b) {
-    if (a == null && b == null) return (0, true);
-    if (a == null) return (1, true);
-    if (b == null) return (-1, true);
-    return (null, false);
+    var res = List<T>.from(shadowThis)
+      ..sort((a, b) =>
+          ExtensionHelper.sorter<K>(UnsupportedError('$K data type is not supported by this order by'))(op(a), op(b)) *
+          (dir == OrderDirection.asc ? 1 : -1));
+    return res;
   }
 }
 
@@ -425,6 +397,21 @@ extension NullableListExtn2<T> on Iterable<T?>? {
       if (element != null) returnList.add(element);
     }
     return returnList;
+  }
+
+  /// Returns a new list that is ordered by the property returned from the operation [op].
+  ///
+  /// Will throw and [UnsupportedError] exception if the data type returned by the [op] is not supported in the sort
+  Iterable<T?> order([OrderDirection dir = OrderDirection.asc]) {
+    if (this == null) return [];
+    var shadowThis = this!;
+    var res = shadowThis.toList()
+      ..sort((a, b) =>
+          ExtensionHelper.sorter<T>(UnsupportedError(
+                  '$T data type is not supported by this order function. Try using orderBy for more complex objects'))(
+              a, b) *
+          (dir == OrderDirection.asc ? 1 : -1));
+    return res;
   }
 }